+++ earth2studio/models/nn/afno_precip.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: Copyright (c) 2024 NVIDIA CORPORATION & AFFILIATES.
+# SPDX-FileCopyrightText: Copyright (c) 2024-2025 NVIDIA CORPORATION & AFFILIATES.
 # SPDX-FileCopyrightText: All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
@@ -14,10 +14,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import modulus
+import physicsnemo
 import torch
 import torch.nn.functional as F
-from modulus.models.afno import AFNO
+from physicsnemo.models.afno import AFNO
 
 
 class PeriodicPad2d(torch.nn.Module):
@@ -33,7 +33,7 @@ class PeriodicPad2d(torch.nn.Module):
         return out
 
 
-class PrecipNet(modulus.Module):
+class PrecipNet(physicsnemo.Module):
     def __init__(
         self,
         inp_shape,