+++ test/io/test_netcdf4.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: Copyright (c) 2024 NVIDIA CORPORATION & AFFILIATES.
+# SPDX-FileCopyrightText: Copyright (c) 2024-2025 NVIDIA CORPORATION & AFFILIATES.
 # SPDX-FileCopyrightText: All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
@@ -17,11 +17,13 @@
 import os
 import tempfile
 from collections import OrderedDict
+from datetime import datetime, timedelta
 
 import netCDF4
 import numpy as np
 import pytest
 import torch
+import xarray
 
 from earth2studio.io import NetCDF4Backend
 from earth2studio.utils.coords import convert_multidim_to_singledim, split_coords
@@ -303,6 +305,44 @@ def test_netcdf4_file(mode: str, device: str, tmp_path: str) -> None:
             nc = NetCDF4Backend(tmp_path / "test.nc", backend_kwargs={"mode": mode})
             nc.add_array(total_coords, array_name, data=dummy)
     nc.close()
+
+
+@pytest.mark.parametrize("device", ["cpu", "cuda:0"])  #
+def test_netcdf4_calendar(device: str, tmp_path: str) -> None:
+    time = [np.datetime64("1958-01-31T00:00:00")]
+    variable = ["t2m", "tcwv"]
+    total_coords = OrderedDict(
+        {
+            "time": np.asarray(time),
+            "variable": np.asarray(variable),
+            "lat": np.linspace(-90, 90, 180),
+            "lon": np.linspace(0, 360, 360, endpoint=False),
+        }
+    )
+
+    # Test File Store
+    nc = NetCDF4Backend(tmp_path / "test.nc", backend_kwargs={"mode": "w"})
+    assert isinstance(nc.root, netCDF4.Dataset)
+
+    shape = tuple([len(values) for values in total_coords.values()])
+    array_name = "fields"
+    dummy = torch.randn(shape, device=device, dtype=torch.float32)
+    nc.add_array(total_coords, array_name, data=dummy)
+    nc.close()
+
+    # Check that the calendar matches with xarray and netCDF4
+    ds = xarray.open_dataset(tmp_path / "test.nc")
+    xarray_time = ds.time[0].values
+
+    # Convert to datetime
+    _unix = np.datetime64(0, "s")
+    _ds = np.timedelta64(1, "s")
+    xarray_datetime = datetime.utcfromtimestamp((xarray_time - _unix) / _ds)
+
+    # Get the time from netCDF4
+    netcdf4_hours = netCDF4.Dataset(tmp_path / "test.nc")["time"][0].data.item()
+    netcdf4_datetime = datetime(1, 1, 1) + timedelta(hours=netcdf4_hours)
+    assert xarray_datetime == netcdf4_datetime
 
 
 @pytest.mark.parametrize(